Start boilerplate for app-level codec
