+++ x/staking/alias.go
@@ -176,6 +176,7 @@ var (
 	NewDescription                     = types.NewDescription
 
 	// variable aliases
+	NewCodec                         = types.NewCodec
 	ModuleCdc                        = types.ModuleCdc
 	LastValidatorPowerKey            = types.LastValidatorPowerKey
 	LastTotalPowerKey                = types.LastTotalPowerKey
@@ -200,6 +201,7 @@ var (
 
 type (
 	Keeper                    = keeper.Keeper
+	Codec                     = types.Codec
 	Commission                = types.Commission
 	CommissionRates           = types.CommissionRates
 	DVPair                    = types.DVPair
